Presets: refactor save/load logic out of PresetManager (#414)
* Presets: refactor save/load logic out of PresetManager

* Add comment
