+++ modules/plugin/chowdsp_presets_v2/Backend/chowdsp_PresetManager.cpp
@@ -2,162 +2,16 @@
 
 namespace chowdsp::presets
 {
-void PresetManager::initializeListeners (ParamHolder& params, ParameterListeners& paramListeners)
+PresetManager::PresetManager (PluginState& state,
+                              juce::AudioProcessor* plugin,
+                              const juce::String& presetFileExtension,
+                              std::vector<juce::RangedAudioParameter*>&& presetAgnosticParams,
+                              PresetTree::InsertionHelper&& insertionHelper)
+    : saverLoader (state, plugin, std::move (presetAgnosticParams)),
+      presetTree (&saverLoader.getPresetState(), std::move (insertionHelper)),
+      presetFileExt (presetFileExtension)
 {
-    params.doForAllParameters (
-        [this, &paramListeners] (auto& param, size_t)
-        {
-            if (isPresetAgnosticParameter (param))
-                return;
-
-            listeners += {
-                paramListeners.addParameterListener (
-                    param,
-                    ParameterListenerThread::MessageThread,
-                    [this]
-                    {
-                        if (! areWeInTheMidstOfAPresetChange)
-                            isPresetDirty.set (true);
-                    })
-            };
-        });
-
-    listeners +=
-        {
-            isPresetDirty.changeBroadcaster.connect (
-                [this]
-                {
-                    if (! areWeInTheMidstOfAPresetChange)
-                        presetDirtyStatusBroadcaster();
-                }),
-            currentPreset.changeBroadcaster.connect (
-                [this]
-                {
-                    if (currentPreset == nullptr)
-                        return;
-
-                    juce::ScopedValueSetter svs { areWeInTheMidstOfAPresetChange, true };
-
-                    loadPresetState (currentPreset->getState()); // @TODO: could this throw?
-                    pluginState.getParameterListeners().updateBroadcastersFromMessageThread();
-                    isPresetDirty.set (false);
-                    presetChangedBroadcaster();
-
-                    if (processor != nullptr)
-                    {
-#if JUCE_VERSION > 0x60007
-                        processor->updateHostDisplay (juce::AudioProcessorListener::ChangeDetails().withProgramChanged (true));
-#else
-                        processor->updateHostDisplay();
-#endif
-                    }
-                })
-        };
-}
-
-void PresetManager::loadPreset (const Preset& preset)
-{
-    struct ChangePresetAction : juce::UndoableAction
-    {
-        explicit ChangePresetAction (PresetManager& mgr, const Preset& preset)
-            : manager (mgr),
-              performPreset (preset),
-              undoPreset (manager.currentPreset->getName(),
-                          manager.currentPreset->getVendor(),
-                          manager.savePresetState(),
-                          manager.currentPreset->getCategory(),
-                          manager.currentPreset->getPresetFile()),
-              previousStateWasDirty (manager.isPresetDirty)
-        {
-        }
-
-        bool perform() override
-        {
-            manager.currentPreset = performPreset;
-            return true;
-        }
-
-        bool undo() override
-        {
-            manager.currentPreset = undoPreset;
-            manager.isPresetDirty.set (previousStateWasDirty);
-            return true;
-        }
-
-        int getSizeInUnits() override { return (int) sizeof (*this); }
-
-        PresetManager& manager;
-        const Preset& performPreset;
-        const Preset undoPreset;
-        const bool previousStateWasDirty = false;
-    };
-
-    pluginState.callOnMainThread (
-        [this, &preset]
-        {
-            if (currentPreset == nullptr || pluginState.undoManager == nullptr)
-            {
-                currentPreset = preset;
-                return;
-            }
-
-            pluginState.undoManager->beginNewTransaction ("Loading preset: " + preset.getName());
-            pluginState.undoManager->perform (std::make_unique<ChangePresetAction> (*this, preset).release());
-        });
-}
-
-bool PresetManager::isPresetAgnosticParameter (const juce::RangedAudioParameter& param) const
-{
-    return std::find_if (
-               presetAgnosticParameters.begin(),
-               presetAgnosticParameters.end(),
-               [&param] (const juce::RangedAudioParameter* presetAgnosticParam)
-               {
-                   return param.paramID == presetAgnosticParam->paramID;
-               })
-           != presetAgnosticParameters.end();
-}
-
-nlohmann::json PresetManager::savePresetState()
-{
-    nlohmann::json state;
-    std::as_const (pluginState)
-        .getParameters()
-        .doForAllParameters (
-            [this, &state] (auto& param, size_t)
-            {
-                if (isPresetAgnosticParameter (param))
-                    return;
-
-                state[param.paramID.toStdString()] = ParameterTypeHelpers::getValue (param);
-            });
-    return state;
-}
-
-void PresetManager::loadPresetState (const nlohmann::json& state)
-{
-    if (currentPreset != nullptr)
-    {
-        const auto newPresetName = currentPreset->getName();
-        juce::Logger::writeToLog ("Loading preset: " + newPresetName);
-    }
-
-    pluginState
-        .getParameters()
-        .doForAllParameters (
-            [this, &state] (auto& param, size_t)
-            {
-                if (isPresetAgnosticParameter (param))
-                    return;
-
-                if (state.contains (param.paramID.toStdString()))
-                {
-                    ParameterTypeHelpers::setValue (state[param.paramID.toStdString()], param);
-                    return;
-                }
-
-                ParameterTypeHelpers::resetParameter (param);
-            });
+    jassert (presetFileExt[0] == '.'); // invalid file extension!
 }
 
 void PresetManager::addPresets (std::vector<Preset>&& presets, bool areFactoryPresets)
@@ -182,7 +36,7 @@ Preset PresetManager::getUserPresetForState (const juce::String& presetName, nlo
 void PresetManager::saveUserPreset (const juce::File& file)
 {
     const auto name = file.getFileNameWithoutExtension();
-    saveUserPreset (file, Preset { name, userPresetsVendor, savePresetState() });
+    saveUserPreset (file, Preset { name, userPresetsVendor, saverLoader.savePresetState() });
 }
 
 void PresetManager::saveUserPreset (const juce::File& file, Preset&& preset) // NOSONAR
